moved thread to model
